# ui/desktop/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "goose-server",
    "description": "An AI agent",
    "contact": {
      "name": "Block",
      "email": "ai-oss-tools@block.xyz"
    },
    "license": {
      "name": "Apache-2.0"
    },
    "version": "1.0.29"
  },
  "paths": {
...
    "/confirm": {
      "post": {
        "tags": [
          "super::routes::reply"
        ],
        "operationId": "confirm_permission",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PermissionConfirmationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Permission action is confirmed",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid secret key"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/security/feedback": {
      "post": {
        "tags": [
          "super::routes::reply"
        ],
        "operationId": "submit_security_feedback",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SecurityFeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Security feedback submitted successfully",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid secret key"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PermissionConfirmationRequest": {
        "type": "object",
        "required": [
          "id",
          "action"
        ],
        "properties": {
          "action": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "principal_type": {
            "$ref": "#/components/schemas/PrincipalType"
          }
        }
      },
...
      "PrincipalType": {
        "type": "string",
        "enum": [
          "Extension",
          "Tool"
        ]
      },
...
      "SecurityFeedbackRequest": {
        "type": "object",
        "required": [
          "note_id",
          "feedback_type"
        ],
        "properties": {
          "feedback_type": {
            "type": "string"
          },
          "note_id": {
            "type": "string"
          },
          "user_comment": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  }
}